Moved most logic from Favorite/WatchLater viewmodels in abstract class
