+++ app/src/main/java/com/sandev/moviesearcher/domain/components_holders/SavedMoviesComponentHolder.kt
@@ -1,0 +1,8 @@
+package com.sandev.moviesearcher.domain.components_holders
+
+import com.sandev.moviesearcher.domain.interactors.MoviesListInteractor
+
+
+interface SavedMoviesComponentHolder {
+    var interactor: MoviesListInteractor
+}
